+++ changelog.md
@@ -20,6 +20,7 @@
 - widget weather.current takes weather values from backend item alternativly to online weather informaion
 - new widget event "exit" allows stopping of functions before page change (e.g. deleting timers)
 - new widget basic.roundslider with flexible design options
+- new widget device.rtrslider for slider control of set temperature and functionality of device.rtr 
 
 ### Other New Features
 - bash script "setpermissions" added for setting file permissions during initial setup. To be called by 'sudo bash setpermissions'